requirements.txt
Django>=5.0
djangorestframework>=3.14
django-filter>=24.0
drf-spectacular>=0.27
django-safedelete>=1.3
python-dotenv>=1.0
requests>=2.28
django-cron>=0.6